Compile in Mac

+++ Rakefile
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require 'os'
 
 # Set HTTPS Proxy for GitHub
 HTTPS_PROXY = ENV['HTTPS_PROXY'] || ENV['https_proxy']
@@ -6,28 +7,51 @@ HTTPS_PROXY = ENV['HTTPS_PROXY'] || ENV['https_proxy']
 # Use musl toolchain
 # Arch Linux: pacman -S musl
 # Debian: apt install musl-dev musl-tools
-CC = 'musl-gcc'
+CC = if OS.mac?
+       'cc'
+     else
+       'musl-gcc'
+     end
 
 BUILD_DIR = 'build'
 C_SOURCE = 'src/main.c'
 
 MRUBY_SOURCE = "#{BUILD_DIR}/mruby"
+
 MRUBY_INCLUDE = "#{MRUBY_SOURCE}/include"
 MRUBY_BUILD = "#{MRUBY_SOURCE}/build"
+
 LIB_MRUBY = "#{MRUBY_BUILD}/host/lib/libmruby.a"
 
 LIBRESSL_SOURCE = "#{BUILD_DIR}/libressl"
-LIBRESSL_INSTALL = '/tmp/libressl'
-LIBRESSL_INCLUDE = "#{LIBRESSL_INSTALL}/usr/local/include/"
-LIBRESSL_LIB = "#{LIBRESSL_INSTALL}/usr/local/lib"
+LIBRESSL_INSTALL = if OS.mac?
+                     `brew --prefix libressl`.strip
+                   else
+                     '/tmp/libressl'
+                   end
+LIBRESSL_INCLUDE = if OS.mac?
+                     "#{LIBRESSL_INSTALL}/include/"
+                   else
+                     "#{LIBRESSL_INSTALL}/usr/local/include/"
+                   end
+LIBRESSL_LIB = if OS.mac?
+                 "#{LIBRESSL_INSTALL}/lib"
+               else
+                 "#{LIBRESSL_INSTALL}/usr/local/lib"
+               end
 LIB_TLS = "#{LIBRESSL_LIB}/libtls.a"
 LIB_SSL = "#{LIBRESSL_LIB}/libssl.a"
 
-MAKE_JOBS = 16
+MAKE_JOBS = OS.cpu_count
 
 EXECUTABLE = 'lynx'
 
 namespace :lynx do
+  desc 'Generate lynx binary which is dynatmic linked to libressl for test only'
+  task mac_compile: 'mruby:mac_libmruby' do
+    sh "#{CC} -std=c99 -I#{MRUBY_INCLUDE} -I#{LIBRESSL_INCLUDE} #{C_SOURCE} #{LIB_MRUBY} #{LIB_SSL} #{LIB_TLS} -o #{BUILD_DIR}/#{EXECUTABLE}"
+  end
+
   desc 'Generate lynx binary which is statically linked against musl'
   task compile: 'mruby:libmruby' do
     sh "#{CC} -std=c99 -I#{MRUBY_INCLUDE} -I#{LIBRESSL_INCLUDE} #{C_SOURCE} #{LIB_MRUBY} #{LIB_SSL} #{LIB_TLS} -lm -static -o #{BUILD_DIR}/#{EXECUTABLE}"
@@ -42,6 +66,14 @@ namespace :lynx do
 end
 
 namespace :mruby do
+  desc 'Build mac version of libmruby.a'
+  task mac_libmruby: %i[mruby:mac_config] do
+    unless File.exist?(LIB_MRUBY.to_s)
+      sh "cd #{BUILD_DIR}/mruby && https_proxy=#{HTTPS_PROXY} rake"
+      sh "cd #{BUILD_DIR}/mruby && https_proxy=#{HTTPS_PROXY} rake test"
+    end
+  end
+
   desc 'Build libmruby.a'
   task libmruby: %i[mruby:config libressl:libssl] do
     unless File.exist?(LIB_MRUBY.to_s)
@@ -54,6 +86,12 @@ namespace :mruby do
   task config: :source do
     sh "rm -fv  #{MRUBY_SOURCE}/build_config/*.rb"
     sh "cp build_config/musl_linux_amd64.rb  #{MRUBY_SOURCE}/build_config/default.rb"
+  end
+
+  desc 'Configure mruby in mac'
+  task mac_config: :source do
+    sh "rm -fv  #{MRUBY_SOURCE}/build_config/*.rb"
+    sh "cp build_config/macos_amd64.rb  #{MRUBY_SOURCE}/build_config/default.rb"
   end
 
   desc 'Fetch mruby source from GitHub'
@@ -100,16 +138,16 @@ namespace :libressl do
 
   desc 'Remove libressl build files'
   task :clean do
-    sh "cd #{LIBRESSL_SOURCE} && make clean"
-    sh "rm -rfv #{LIBRESSL_INSTALL}"
-    sh "rm -rfv #{LIBRESSL_INSTALL}."
+    sh "cd #{LIBRESSL_SOURCE} && make clean" unless OS.mac?
+    sh "rm -rfv #{LIBRESSL_INSTALL}" unless OS.mac?
+    sh "rm -rfv #{LIBRESSL_INSTALL}." unless OS.mac?
   end
 
   desc 'Remove libressl source and build files'
   task :clobber do
-    sh "rm -rfv #{LIBRESSL_SOURCE}"
-    sh "rm -rfv #{LIBRESSL_INSTALL}"
-    sh "rm -rfv #{LIBRESSL_INSTALL}."
+    sh "rm -rfv #{LIBRESSL_SOURCE}" unless OS.mac?
+    sh "rm -rfv #{LIBRESSL_INSTALL}" unless OS.mac?
+    sh "rm -rfv #{LIBRESSL_INSTALL}." unless OS.mac?
   end
 end
 
@@ -120,7 +158,7 @@ end
 
 namespace :run do
   desc 'Build release'
-  task build: %i[lynx:compile]
+  task build: (OS.mac? ? %i[lynx:mac_compile] : %i[lynx:compile])
 
   desc 'Remove build files'
   task clean: %i[lynx:clean mruby:clean libressl:clean]
